requirements.txt
numpy
pandas
pyyaml
matplotlib

autograd
scipy
sympy
scikit-learn

python-dotenv
requests
debugpy